# internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lurkarr API
  description: >-
    REST API for Lurkarr — an automated media management tool that integrates
    with Sonarr, Radarr, Lidarr, Readarr, Whisparr, Eros, Prowlarr, SABnzbd,
    and Seerr.
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/user/username:
    post:
      operationId: updateUsername
      tags: [User]
      summary: Update username
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/password:
    post:
      operationId: updatePassword
      tags: [User]
      summary: Update password
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Instances ─────────────────────────────────────────────────────────
  /api/instances/{app}:
    get:
      operationId: listInstances
      tags: [Instances]
      summary: List instances for an app type
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AppTypePath'
      responses:
        '200':
          description: List of instances (API keys masked)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppInstance'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      operationId: createInstance
      tags: [Instances]
      summary: Create a new instance
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AppTypePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceRequest'
      responses:
        '201':
          description: Instance created (API key masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstance'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/instances/{id}:
    put:
      operationId: updateInstance
      tags: [Instances]
      summary: Update an instance
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UUIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstanceRequest'
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteInstance
      tags: [Instances]
      summary: Delete an instance
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UUIDPath'
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/settings/general:
    get:
      operationId: getGeneralSettings
      tags: [Settings]
      summary: Get general settings
      description: Secret key is masked in the response.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: General settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralSettings'
    put:
      operationId: updateGeneralSettings
      tags: [Settings]
      summary: Update general settings
      description: Secret key is always preserved server-side and cannot be changed via API.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneralSettings'
      responses:
        '200':
          description: Updated settings (secret key masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lurkarr_session
  parameters:
    AppTypePath:
      name: app
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/AppType'
    UUIDPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    StatusOk:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                const: ok
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    AppType:
      type: string
      enum: [sonarr, radarr, lidarr, readarr, whisparr, eros, prowlarr]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AppInstance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        app_type:
          $ref: '#/components/schemas/AppType'
        name:
          type: string
        api_url:
          type: string
        api_key:
          type: string
          description: Masked in responses (****XXXX)
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateInstanceRequest:
      type: object
      required: [name, api_url, api_key]
      properties:
        name:
          type: string
        api_url:
          type: string
        api_key:
          type: string
    UpdateInstanceRequest:
      type: object
      properties:
        name:
          type: string
        api_url:
          type: string
        api_key:
          type: string
          description: Send masked value (****...) to preserve existing key
        enabled:
          type: boolean
...
    GeneralSettings:
      type: object
      properties:
        secret_key:
          type: string
          description: Always masked as "****" in responses. Cannot be changed via API.
        proxy_auth_bypass:
          type: boolean
        ssl_verify:
          type: boolean
        api_timeout:
          type: integer
          minimum: 1
        stateful_reset_hours:
          type: integer
          minimum: 0
        command_wait_delay:
          type: integer
          minimum: 0
        command_wait_attempts:
          type: integer
          minimum: 0
        min_download_queue_size:
          type: integer
          minimum: 0
...